+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- Drop Final annotation from Enum in semantic conventions
+  ([#4085](https://github.com/open-telemetry/opentelemetry-python/pull/4085))
+
 ## Version 1.26.0/0.47b0 (2024-07-25)
 
 - Standardizing timeout calculation in measurement consumer collect to nanoseconds